Feature Selection

Function to conduct feature selection on snRNA-seq and spatial transcriptomics data and calculate the difference of deviance and rank with and without the selected batch effect.

Usage

featureSelect(input, batch_effect = NULL, VGs = NULL)

Arguments

input

SpatialExperiment or SingleCellExperiment: Input data which can be either SpatialExperiment or SingleCellExperiment. It is assumed to have an assays slot containing counts assay for devianceFeatureSelection() to successfully operate and calculate the deviance and rank. The logcounts is strongly recommended to be included in assays. Also, the input data is assumed to set up rownames(input) as genes and gene_name in rowData(input).

batch_effect

character: Either batch on slide or subject based on what metadata is available within input data. The name of slide or subject within each input object should be specified based on different scenarios.

VGs

character: Highly Variable Genes (HVGs) for SingleCelleExperiment object or Spatially Variable Genes (SVGs) for SpatialExperiment object. If it is a data frame, it is assumed to contain one column of identified variable genes with column name as "gene_name".

Value

If the input was provided as a SpatialExperiment or SingleCellExperiment object, the output values are returned as a data frame containing the deviance and rank with and without the bacth effect, the corresponding difference, the corresponding nSD, gene, gene name, and whether the gene is outlier defined by the chosen deviance and rank cutoff.

Examples

library(SummarizedExperiment)
data(spe)
spe_svgs <- spe
# the sample dataset here is the SpatialExperiment object containing only
# spatially variable genes
SVGs <- rowData(spe)$gene_name

set.seed(123)
batch_df <- featureSelect(spe_svgs, batch_effect = "sample_id", VGs = SVGs)
#> Step 1: Running feature selection without batch...
#> Step 2: Running feature selection with batch...
#> Step 3: Calculating deviance and rank difference...