Find Bias Genes

Function to run 'BiasDetect' method for spatially variable genes (SVGs) identification on spatial-resolved transcriptomics (SRT) data with some types of batch effect.

Usage

biasDetect(
  input,
  batch_effect = NULL,
  VGs = NULL,
  nSD_dev = 5,
  nSD_rank = 5,
  visual = FALSE
)

Arguments

input

SpatialExperiment or SingleCellExperiment: Input data which can be either SpatialExperiment or SingleCellExperiment. It is assumed to have an assays slot containing counts assay for devianceFeatureSelection() to successfully operate and calculate the deviance and rank. The logcounts is strongly recommended to be included in assays. Also, the input data is assumed to set up rownames(input) as genes and gene_name in rowData(input).

batch_effect

character: either batch on slide or subject based on what metadata is available within input data. The name of slide or subject within each input object should be specified based on different scenarios.

VGs

character: highly variable genes (HVGs) for SingleCelleExperiment object or spatially variable genes (SVGs) for SpatialExperiment object. If it is a data frame, it is assumed to contain one column of identified variable genes with column name as "gene_name".

nSD_dev

integer: number of standard deviation (nSD) on deviance difference. The default value is 5.

nSD_rank

integer: number of standard deviation (nSD) on rank difference. The default value is 5.

visual

logical: Whether to display the detected bias genes by visualizations of deviance and rank. Default = FALSE to return the data frame. If it is TRUE, the returned format will be two parallel plots presenting bias genes based on both nSD_dev and nSD_rank.

Value

If the input was provided as a SpatialExperiment or SingleCellExperiment object, the output values are returned as either a data frame or two parallel plots to present the identified bias genes.

Examples

# \donttest{
set.seed(123)
spe <- load("inst/data/spe_norm.Rdata")
#> Warning: cannot open compressed file 'inst/data/spe_norm.Rdata', probable reason 'No such file or directory'
#> Error in readChar(con, 5L, useBytes = TRUE): cannot open the connection
libd <- read.csv("inst/data/libd-all_nnSVG_p-05-features-df.csv/")
#> Warning: cannot open file 'inst/data/libd-all_nnSVG_p-05-features-df.csv/': No such file or directory
#> Error in file(file, "rt"): cannot open the connection
SVGs <- libd$gene_name
#> Error in eval(expr, envir, enclos): object 'libd' not found
biasDetect(spe, batch_effect = "brain", VGs = SVGs, nSD_dev = 5,
        nSD_rank = 6, visual = FALSE)
#> Error in eval(expr, envir, enclos): object 'spe' not found
 biasDetect(spe, batch_effect = "brain", VGs = SVGs, nSD_dev = 5,
        nSD_rank = 6, visual = TRUE)
#> Error in eval(expr, envir, enclos): object 'spe' not found
# }